Clean up leftover comments after extraction
