+++ app/Http/Controllers/EventsController.php
@@ -1389,752 +1389,6 @@ class EventsController extends Controller
         ];
     }
 
-    /**
-     * Endpoint to post a single event to Instagram.
-     */
-    public function postToInstagram(int $id, Instagram $instagram): RedirectResponse
-    {
-        // load the event
-        if (!$event = Event::find($id)) {
-            flash()->error('Error', 'No such event');
-
-            return back();
-        }
-
-        // get the instagram account
-        if (!$instagram->getIgUserId()) {
-            flash()->error('Error', 'You must have an Instagram user account linked to post to Instagram.');
-
-            return back();
-        }
-
-        // get the instagram page access token
-        if (!$instagram->getPageAccessToken()) {
-            flash()->error('Error', 'You must have an Instagram page linked to post to Instagram.');
-
-            return back();
-        }
-
-        // get the image URL
-        $photo = $event->getPrimaryPhoto();
-
-        if (!$photo) {
-            flash()->error('Error', 'You must have an photo to extract the image to post to Instagram');
-
-            return back();
-        }
-
-        $imageUrl = Storage::disk('external')->url($photo->getStoragePath());
-
-        if (!$imageUrl) {
-            flash()->error('Error', 'You must have an image url to post to Instagram');
-
-            return back();
-        }
-
-        // get the instagram caption
-        $caption = urlEncode($event->getInstagramFormat());
-
-        if (!$caption) {
-            flash()->error('Error', 'You must have an Instagram caption linked to post to Instagram.');
-
-            return back();
-        }
-
-        // make the instagram api calls
-
-        // upload the image
-        try {
-            $igContainerId = $instagram->uploadPhoto($imageUrl, $caption);
-        } catch (Exception $e) {
-            flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-
-            return back();
-        }
-
-        // check the container status every 5 seconds until status_code is FINISHED
-        if ($instagram->checkStatus($igContainerId) === false) {
-            flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-
-            return back();
-        }
-
-        // pubish the image
-        $result = $instagram->publishMedia($igContainerId);
-        if ($result === false) {
-            flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-
-            return back();
-        }
-
-        // log the post to instagram
-        Activity::log($event, $this->user, 16);
-
-        // post was successful
-        flash()->success('Success', 'Successfully published to Instagram, returned id: '.$result);
-
-        return back();
-    }
-
-
-        
-    /**
-     * Endpoint to post a single event as a carousel to Instagram.
-     */
-    public function postCarouselToInstagram(int $id, Instagram $instagram): RedirectResponse
-    {
-        // load the event
-        if (!$event = Event::find($id)) {
-            flash()->error('Error', 'No such event');
-
-            return back();
-        }
-
-        // get the instagram account
-        if (!$instagram->getIgUserId()) {
-            flash()->error('Error', 'You must have an Instagram user account linked to post to Instagram.');
-
-            return back();
-        }
-
-        // get the instagram page access token
-        if (!$instagram->getPageAccessToken()) {
-            flash()->error('Error', 'You must have an Instagram page linked to post to Instagram.');
-
-            return back();
-        }
-
-        // get the image URL
-        $photo = $event->getPrimaryPhoto();
-
-        if (!$photo) {
-            flash()->error('Error', 'You must have an photo to extract the image to post to Instagram');
-
-            return back();
-        }
-
-        $imageUrl = Storage::disk('external')->url($photo->getStoragePath());
-
-        if (!$imageUrl) {
-            flash()->error('Error', 'You must have an image url to post to Instagram');
-
-            return back();
-        }
-
-        // get the instagram caption
-        $caption = urlEncode($event->getInstagramFormat());
-        $caption = $event->getInstagramFormat();
-
-        // make the instagram api calls
-        $igContainerIds = [];
-
-        // upload the image
-        try {
-            $id = $instagram->uploadCarouselPhoto($imageUrl);
-            $igContainerIds[] = $id;
-        } catch (Exception $e) {
-            flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-            Log::info('Carousel photo error: '. $e->getMessage());
-            return back();
-        }
-
-        Log::info('Carousel photo uploaded: '.$id);
-
-        // add other photos related to the event to the carousel
-        foreach ($event->getOtherPhotos() as $otherPhotos) {
-            $imageUrl = Storage::disk('external')->url($otherPhotos->getStoragePath());
-
-            if (!$imageUrl) {
-                flash()->error('Error', 'You must have an image url to post to Instagram');
-                Log::info('No image url found for event: '.$event->id);
-                continue;
-            }
-
-            try {
-                $igContainerId = $instagram->uploadCarouselPhoto($imageUrl);
-            } catch (Exception $e) {
-                flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-                Log::info('Carousel photo error: '. $e->getMessage());
-                return back();
-            }
-
-            // check the container status every 5 seconds until status_code is FINISHED
-            if ($instagram->checkStatus($igContainerId) === false) {
-                flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-                Log::info('Error checking status of carousel photo');
-                return back();
-            }
-
-            $igContainerIds[] = $igContainerId;
-            Log::info('Added container id: '.$igContainerId);
-
-            Log::info('Carousel photo uploaded: '.$id);
-        }
-
-        // only do this if there are any other related photos
-        foreach ($event->entities as $entity) {
-            foreach ($entity->photos as $photo) {
-                if ($photo->is_primary) {
-                    // process the photo
-                    $imageUrl = Storage::disk('external')->url($photo->getStoragePath());
-                    $images[] = $imageUrl;
-
-                    if (!$imageUrl) {
-                        flash()->error('Error', 'You must have an image url to post to Instagram');
-            
-                        return back();
-                    }
-
-                    // make the instagram api calls
-                    // upload the image
-                    try {
-                        $igContainerId = $instagram->uploadCarouselPhoto($imageUrl);
-                    } catch (Exception $e) {
-                        flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-                        Log::info('Error uploading carousel photo');
-                        return back();
-                    }
-
-                    // check the container status every 5 seconds until status_code is FINISHED
-                    if ($instagram->checkStatus($igContainerId) === false) {
-                        flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-                        Log::info('Error checking status of carousel photo');
-                        return back();
-                    }
-
-                    $igContainerIds[] = $igContainerId;
-                    Log::info('Added container id: '.$igContainerId);
-                }
-            }
-        }
-
-        // create a carousel container
-        try {
-            $igCarouselId = $instagram->createCarousel($igContainerIds, $caption);
-        } catch (Exception $e) {
-            flash()->error('Error', 'There was an error posting carousel to Instagram.  Please try again.');
-            Log::info('Error creating carousel');
-            return back();
-        }
-
-        // check the container status every 5 seconds until status_code is FINISHED
-        if ($instagram->checkStatus($igCarouselId) === false) {
-            flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-
-            return back();
-        }
-
-        // pubish the image
-        $result = $instagram->publishMedia($igCarouselId);
-        if ($result === false) {
-            flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-
-            return back();
-        }
-        Log::info('Carousel published: '.$igCarouselId);
-
-        // log the post to instagram
-        Activity::log($event, $this->user, 16);
-
-        // post was successful
-        flash()->success('Success', 'Successfully published to Instagram, returned id: '.$result);
-
-        return back();
-    }
-
-
-    /**
-     * Endpoint to post an event as a STORY to Instagram.
-     */
-    public function postStoryToInstagram(int $id, Instagram $instagram): RedirectResponse
-    {
-        // load the event
-        if (!$event = Event::find($id)) {
-            flash()->error('Error', 'No such event');
-
-            return back();
-        }
-
-        // get the instagram account
-        if (!$instagram->getIgUserId()) {
-            flash()->error('Error', 'You must have an Instagram user account linked to post to Instagram.');
-
-            return back();
-        }
-
-        // get the instagram page access token
-        if (!$instagram->getPageAccessToken()) {
-            flash()->error('Error', 'You must have an Instagram page linked to post to Instagram.');
-
-            return back();
-        }
-
-        // get the image URL
-        $photo = $event->getPrimaryPhoto();
-
-        if (!$photo) {
-            flash()->error('Error', 'You must have an photo to extract the image to post to Instagram');
-
-            return back();
-        }
-
-        $imageUrl = Storage::disk('external')->url($photo->getStoragePath());
-
-        if (!$imageUrl) {
-            flash()->error('Error', 'You must have an image url to post to Instagram');
-
-            return back();
-        }
-
-        // get the instagram caption
-        $caption = urlEncode($event->getInstagramFormat());
-        $caption = $event->getInstagramFormat();
-
-        // make the instagram api calls
-        $igContainerIds = [];
-
-        // upload the image
-        try {
-            $id = $instagram->uploadCarouselPhoto($imageUrl);
-            $igContainerIds[] = $id;
-        } catch (Exception $e) {
-            flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-            Log::info('Carousel photo error: '. $e->getMessage());
-            return back();
-        }
-
-        Log::info('Carousel photo uploaded: '.$id);
-
-        // add other photos related to the event to the carousel
-        foreach ($event->getOtherPhotos() as $otherPhotos) {
-            $imageUrl = Storage::disk('external')->url($otherPhotos->getStoragePath());
-
-            if (!$imageUrl) {
-                flash()->error('Error', 'You must have an image url to post to Instagram');
-                Log::info('No image url found for event: '.$event->id);
-                continue;
-            }
-
-            try {
-                $igContainerId = $instagram->uploadCarouselPhoto($imageUrl);
-            } catch (Exception $e) {
-                flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-                Log::info('Carousel photo error: '. $e->getMessage());
-                return back();
-            }
-
-            // check the container status every 5 seconds until status_code is FINISHED
-            if ($instagram->checkStatus($igContainerId) === false) {
-                flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-                Log::info('Error checking status of carousel photo');
-                return back();
-            }
-
-            $igContainerIds[] = $igContainerId;
-            Log::info('Added container id: '.$igContainerId);
-
-            Log::info('Carousel photo uploaded: '.$id);
-        }
-
-        // only do this if there are any other related photos
-        foreach ($event->entities as $entity) {
-            foreach ($entity->photos as $photo) {
-                if ($photo->is_primary) {
-                    // process the photo
-                    $imageUrl = Storage::disk('external')->url($photo->getStoragePath());
-                    $images[] = $imageUrl;
-
-                    if (!$imageUrl) {
-                        flash()->error('Error', 'You must have an image url to post to Instagram');
-            
-                        return back();
-                    }
-
-                    // make the instagram api calls
-                    // upload the image
-                    try {
-                        $igContainerId = $instagram->uploadCarouselPhoto($imageUrl);
-                    } catch (Exception $e) {
-                        flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-                        Log::info('Error uploading carousel photo');
-                        return back();
-                    }
-
-                    // check the container status every 5 seconds until status_code is FINISHED
-                    if ($instagram->checkStatus($igContainerId) === false) {
-                        flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-                        Log::info('Error checking status of carousel photo');
-                        return back();
-                    }
-
-                    $igContainerIds[] = $igContainerId;
-                    Log::info('Added container id: '.$igContainerId);
-                }
-            }
-        }
-
-        // create a carousel container
-        try {
-            $igCarouselId = $instagram->createCarousel($igContainerIds, $caption);
-        } catch (Exception $e) {
-            flash()->error('Error', 'There was an error posting carousel to Instagram.  Please try again.');
-            Log::info('Error creating carousel');
-            return back();
-        }
-
-        // check the container status every 5 seconds until status_code is FINISHED
-        if ($instagram->checkStatus($igCarouselId) === false) {
-            flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-
-            return back();
-        }
-
-        // pubish the image
-        $result = $instagram->publishStoryMedia($igCarouselId);
-        if ($result === false) {
-            flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-
-            return back();
-        }
-        Log::info('Carousel published: '.$igCarouselId);
-
-        // log the post to instagram
-        Activity::log($event, $this->user, 16);
-
-        // post was successful
-        flash()->success('Success', 'Successfully published to Instagram, returned id: '.$result);
-
-        return back();
-    }
-
-
-    /**
-     * Endpoint to post a week's events event to Instagram.
-     */
-    public function postWeekToInstagram(Instagram $instagram, ImageHandler $imageHandler): RedirectResponse
-    {
-        // load the first 9 events of the week
-        $events = Event::where('start_at', '>=', Carbon::now()->startOfWeek())
-            ->where('start_at', '<=', Carbon::now()->endOfWeek())
-            ->orderBy('start_at', 'ASC')
-            ->limit(9)
-            ->get();
-
-        // get the first image to post
-        $coverFileName = 'week-image.jpg';
-        $coverImage = $imageHandler->generateCoverImage($coverFileName);
-
-        if (!$coverImage) {
-            flash()->error('Error', 'You must have a base image to extract the image to make a week post to Instagram');
-
-            return back();
-        }
-
-        // save the file in Storage
-        $coverPath = Storage::disk('external')->putFileAs('photos', new HttpFile($coverImage->basePath()), $coverFileName, 'public');
-        $coverImageUrl = Storage::disk('external')->url($coverPath);
-
-        // create an array of images to post
-        $images[] = $coverImageUrl;
-        $igContainerIds = [];
-
-        Log::info('Cover image URL: '.$coverImageUrl);
-
-        // create a string of event data for the caption
-        $caption = "Events for the upcoming week...\n";
-
-        // get the instagram account
-        if (!$instagram->getIgUserId()) {
-            flash()->error('Error', 'You must have an Instagram user account linked to post to Instagram.');
-
-            return back();
-        }
-
-        // get the instagram page access token
-        if (!$instagram->getPageAccessToken()) {
-            flash()->error('Error', 'You must have an Instagram page linked to post to Instagram.');
-
-            return back();
-        }
-
-        // upload the cover image
-        try {
-            $id = $instagram->uploadCarouselPhoto($coverImageUrl);
-            $igContainerIds[] = $id;
-        } catch (Exception $e) {
-            flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-            Log::info('Carousel photo error: '. $e->getMessage());
-            return back();
-        }
-
-        Log::info('Carousel photo uploaded: '.$id);
-
-        // get info from the events
-        foreach ($events as $event) {
-            // get the image URL
-            $photo = $event->getPrimaryPhoto();
-
-            if (!$photo) {
-                flash()->error('Error', 'You must have an photo to extract the image to post to Instagram');
-                Log::info('No photo found for event: '.$event->id);
-                continue;
-            }
-
-            $imageUrl = Storage::disk('external')->url($photo->getStoragePath());
-            $images[] = $imageUrl;
-
-            if (!$imageUrl) {
-                flash()->error('Error', 'You must have an image url to post to Instagram');
-                Log::info('No image url found for event: '.$event->id);
-                continue;
-            }
-    
-            // get the instagram caption
-            $caption .= $event->getInstagramFormat()."\n\n";
-
-            if (!$caption) {
-                flash()->error('Error', 'You must have an Instagram caption linked to post to Instagram.');
-                Log::info('No caption found for event: '.$event->id);
-                continue;
-            }
-
-            // make the instagram api calls
-            // upload the image
-            try {
-                $igContainerId = $instagram->uploadCarouselPhoto($imageUrl);
-            } catch (Exception $e) {
-                flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-                Log::info('Error uploading carousel photo');
-                return back();
-            }
-
-            // check the container status every 5 seconds until status_code is FINISHED
-            if ($instagram->checkStatus($igContainerId) === false) {
-                flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-                Log::info('Error checking status of carousel photo');
-                return back();
-            }
-
-            $igContainerIds[] = $igContainerId;
-            Log::info('Added container id: '.$igContainerId);
-        }
-
-        Log::info('Looped through events');
-
-        Log::info('Caption:'.$caption);
-        Log::info('IG Container Ids:'.json_encode($igContainerIds));
-
-        // create a carousel container
-        try {
-            $igContainerId = $instagram->createCarousel($igContainerIds, $caption);
-        } catch (Exception $e) {
-            flash()->error('Error', 'There was an error posting carousel to Instagram.  Please try again.');
-            Log::info('Error creating carousel');
-            return back();
-        }
-
-        // check the container status every 5 seconds until status_code is FINISHED
-        if ($instagram->checkStatus($igContainerId) === false) {
-            flash()->error('Error', 'There was an error posting carousel to Instagram.  Please try again.');
-            Log::info('Error checking status');
-            return back();
-        }
-
-        // publish the carousel
-        $result = $instagram->publishMedia($igContainerId);
-        if ($result === false) {
-            flash()->error('Error', 'There was an error posting to Instagram.  Please try again.');
-            Log::info('Error publishing media');
-            return back();
-        }
-
-        // post was successful
-        flash()->success('Success', 'Successfully published to Instagram, returned id: '.$result);
-
-        return back();
-    }
-
-
-    protected function checkBlackList(?Event $event): bool
-    {
-        $blacklist = false;
-
-        // blacklist events that have venues in the blacklist
-        if (!empty($event->venue_id)) {
-            if ($blacklistConfig = config('app.spider_blacklist')) {
-                $blacklistArray = explode(',', $blacklistConfig);
-                foreach ($blacklistArray as $item) {
-                    if (strtolower($item) == strtolower($event->venue->name)) {
-                        $blacklist = true;
-                    }
-                }
-            }
-        }
-
-        return $blacklist;
-    }
-
-    public function show(?Event $event, EmbedExtractor $embedExtractor): string
-    {
-        if (!$event) {
-            abort(404);
-        }
-        $embeds = $embedExtractor->getEmbedsForEvent($event);
-
-        $thread = Thread::where('event_id', '=', $event->id)->first();
-
-        // check blacklist status
-        $blacklist = $this->checkBlackList($event);
-
-        // // extract all the links from the event body and convert into embeds
-        // $embedExtractor->setLayout("small");
-        // $embeds = $embedExtractor->getEmbedsForEvent($event);
-        // $embeds = [];
-
-        return view('events.show', compact('event', 'embeds'))->with(['thread' => $thread, 'blacklist' => $blacklist])->render();
-    }
-
-
-    /**
-     * Load the embeds and add to the UI
-     *
-     * @throws \Throwable
-     */
-    public function loadEmbeds(int $id, EmbedExtractor $embedExtractor, Request $request): RedirectResponse | array
-    {
-        // load the event
-        if (!$event = Event::with('entities.links')->find($id)) {
-            flash()->error('Error', 'No such event');
-
-            return back();
-        }
-
-        // extract all the links from the event body and convert into embeds
-        $embedExtractor->setLayout("medium");
-        $embeds = $embedExtractor->getEmbedsForEvent($event);
-
-        // handle the request if ajax
-        if ($request->ajax()) {
-            return [
-                'Message' => 'Added embeds to event page.',
-                'Success' => view('embeds.playlist')
-                    ->with(compact('embeds'))
-                    ->render(),
-            ];
-        }
-        flash()->success('Error', 'You cannot load embeds directly');
-
-        return back();
-    }
-
-    /**
-     * Load the embeds and add to the UI
-     *
-     * @throws \Throwable
-     */
-    public function loadMinimalEmbeds(int $id, EmbedExtractor $embedExtractor, Request $request): RedirectResponse | array
-    {
-        // load the event
-        if (!$event = Event::with('entities.links')->find($id)) {
-            flash()->error('Error', 'No such event');
-
-            return back();
-        }
-
-        // extract all the links from the event body and convert into embeds
-        $embedExtractor->setLayout("small");
-        $embeds = $embedExtractor->getEmbedsForEvent($event);
-
-        // handle the request if ajax
-        if ($request->ajax()) {
-            return [
-                'Message' => 'Added embeds to event page.',
-                'Success' => view('embeds.minimal-playlist')
-                    ->with(compact('embeds'))
-                    ->render(),
-            ];
-        }
-        flash()->success('Error', 'You cannot load embeds directly');
-
-        return back();
-    }
-
-    public function store(EventRequest $request, Event $event, ImageHandler $imageHandler): RedirectResponse
-    {
-        $msg = '';
-
-        $input = $request->all();
-
-        // transform the slug passed in the request
-        $input['slug'] = Str::slug($request->input('slug', '-'));
-
-        // validation happening in EventRequest->rules
-        $tagArray = $request->input('tag_list', []);
-        $syncArray = [];
-
-        // check the elements in the tag list, and if any don't match, add the tag
-        foreach ($tagArray as $key => $tag) {
-            if (!is_numeric($tag) || !$newTag = Tag::find($tag)) {
-                $newTag = new Tag();
-                $newTag->name = ucwords(strtolower($tag));
-                $newTag->slug = Str::slug($tag);
-                $newTag->tag_type_id = 1;
-                $newTag->save();
-
-                // log adding of new tag
-                Activity::log($newTag, $this->user, 1);
-
-                $syncArray[] = $newTag->id;
-
-                $msg .= ' Added tag '.$tag.'.';
-            } else {
-                $syncArray[] = $newTag->id;
-            }
-        }
-
-        $event = $event->create($input);
-
-        $event->tags()->attach($syncArray);
-        $event->entities()->attach($request->input('entity_list'));
-
-        // also attach the venue and promoter if set
-        if ($request->input('venue_id')) {
-            $event->entities()->syncWithoutDetaching($request->input('venue_id'));
-        }
-
-        if ($request->input('promoter_id')) {
-            $event->entities()->syncWithoutDetaching($request->input('promoter_id'));
-        }
-
-        // add to activity log
-        Activity::log($event, $this->user, 1);
-
-        // TODO figure out why this is failing after Laravel 9 upgrade - doesnt seem to be doing anything, so no big deal..
-        // dispatch notifications that the event was created
-        EventCreated::dispatch($event);
-
-        flash()->success('Success', 'Your event has been created');
-
-        $photo = $event->getPrimaryPhoto();
-
-        // make a call to notify all users who are following any of the tags/keywords if the event starts in the future
-        if ($event->start_at >= Carbon::now()) {
-            // only do the notification if there is a photo
-            if ($photo !== null) {
-                $this->notifyFollowing($event);
-            }
-        }
-
-        // add a twitter notification if the user is admin
-        if (Auth::user()->hasGroup('super_admin') && config('app.twitter_consumer_key') !== '999') {
-            // only tweet if there is a primary photo
-            if ($photo !== null) {
-                $event->notify(new EventPublished());
-            }
-        }
-
-        return redirect()->route('events.show', compact('event'));
-    }
 
     protected function notifyFollowing(Event $event): void
     {
@@ -2216,148 +1470,6 @@ class EventsController extends Controller
                 $newTag->slug = Str::slug($tag);
                 $newTag->tag_type_id = 1;
                 $newTag->save();
-
-                // log adding of new tag
-                Activity::log($newTag, auth()->user(), 1);
-
-                $syncArray[strtolower($tag)] = $newTag->id;
-
-                $msg .= ' Added tag '.$tag.'.';
-            } else {
-                $syncArray[$key] = $tag;
-            }
-        }
-
-        $event->tags()->sync($syncArray);
-        $event->entities()->sync($request->input('entity_list', []));
-
-        // add to activity log
-        Activity::log($event, $this->user, 2);
-        EventUpdated::dispatch($event);
-
-        flash()->success('Success', 'Your event has been updated');
-
-        return redirect()->route('events.show', compact('event'));
-    }
-
-    protected function unauthorized(EventRequest $request): RedirectResponse | Response
-    {
-        if ($request->ajax()) {
-            return response(['message' => 'No way.'], 403);
-        }
-
-        Session::flash('flash_message', 'Not authorized');
-
-        return redirect('/');
-    }
-
-    public function destroy(Event $event): RedirectResponse
-    {
-        // add to activity log
-        Activity::log($event, auth()->user(), 3);
-
-        $event->delete();
-
-        flash()->success('Success', 'Your event has been deleted!');
-
-        return redirect('events');
-    }
-
-    /**
-     * Tweet this event.
-     *
-     * @throws \Throwable
-     */
-    public function tweet(int $id): RedirectResponse
-    {
-        // check if there is a logged in user
-        if (!$this->user) {
-            flash()->error('Error', 'No user is logged in.');
-
-            return back();
-        }
-
-        if (!$event = Event::find($id)) {
-            flash()->error('Error', 'No such event');
-
-            return back();
-        }
-
-        // add a twitter notification
-        $event->notify(new EventPublished());
-
-        // unlink the temp file
-        if ($photo = $event->getPrimaryPhoto()) {
-            unlink(storage_path().'/app/public/photos/temp/'.$photo->name);
-        };
-
-        Log::info('User '.$id.' tweeted '.$event->name);
-
-        flash()->success('Success', 'You tweeted the event - '.$event->name);
-
-        return back();
-    }
-
-    /**
-     * Mark user as attending the event.
-     *
-     * @throws \Throwable
-     */
-    public function attend(int $id, Request $request): RedirectResponse | array
-    {
-        // check if there is a logged in user
-        if (!$this->user) {
-            flash()->error('Error', 'No user is logged in.');
-
-            return back();
-        }
-
-        if (!$event = Event::find($id)) {
-            flash()->error('Error', 'No such event');
-
-            return back();
-        }
-
-        // add the attending response
-        $response = new EventResponse();
-        $response->event()->associate($id);
-        $response->user()->associate($this->user);
-        $response->responseType()->associate(ResponseType::find(1)); // 1 = Attending, 2 = Interested, 3 = Uninterested, 4 = Cannot Attend
-        $response->save();
-
-        // add to activity log
-        Activity::log($event, $this->user, 6);
-
-        // handle the request if ajax
-        if ($request->ajax()) {
-            return [
-                'Message' => 'You are now attending the event - '.$event->name,
-                'Success' => view('events.single')
-                    ->with(compact('event'))
-                    ->with('month', '')
-                    ->render(),
-            ];
-        }
-        flash()->success('Success', 'You are now attending the event - '.$event->name);
-
-        return back();
-    }
-
-    /**
-     * Mark user as unattending the event.
-     *
-     * @throws \Throwable
-     */
-    public function unattend(int $id, Request $request): RedirectResponse | array
-    {
-        // check if there is a logged in user
-        if (!$this->user) {
-            flash()->error('Error', 'No user is logged in.');
-
-            return back();
-        }
-
-        if (!$event = Event::find($id)) {
             flash()->error('Error', 'No such event');
 
             return back();
